Impl Commit

+++ src/main/kotlin/jp/ac/titech/cs/se/refactorhub/models/User.kt
@@ -8,7 +8,7 @@ import javax.persistence.*
 data class User(
     @Id
     @Column(name = "id", nullable = false)
-    val id: Long = 0,
+    val id: Int = 0,
 
     @Column(name = "name", nullable = false, unique = true)
     var name: String = ""
@@ -16,7 +16,7 @@ data class User(
     data class Principal(
         val attributes: Map<String, Any>
     ) : AuthenticatedPrincipal {
-        val id: Long get() = attributes["id"] as Long
+        val id: Int get() = attributes["id"] as Int
         override fun getName(): String = attributes["login"] as String
     }
 }
